# Faster keypoint extraction & post-processing