… scripts/checkpoint_inference.py
… test/test_heatmap.py