Refactored keypoint extraction to use torch.maxpool and torch.topk instead of scipy.

Also operates on batches of channels instead of individual channels

Depending on size of heatmaps, batch_size and # keypoints, speedups are between 5x and 70x:
```
benchmarking with  batch_size: 1, 2 channels and 10 keypoints per channel

5.231 ms per iter for scipy method with heatmap size (256, 256) 
0.653 ms per iter for torch method with heatmap size (256, 256) 

8.732 ms per iter for scipy method with heatmap size (512, 256) 
0.520 ms per iter for torch method with heatmap size (512, 256) 

13.223 ms per iter for scipy method with heatmap size (512, 512) 
0.524 ms per iter for torch method with heatmap size (512, 512) 

109.976 ms per iter for scipy method with heatmap size (1920, 1080) 
1.152 ms per iter for torch method with heatmap size (1920, 1080) 
``` 

(results from `scripts/benchmark_heatmaps.py`)

This improves both inference speed and validation/test speed during training.